Correction to swissN_first_round fix in 5cc53e05
We should only use swissN_first_round() if no player in the relevant
division has played any games yet. This is not the same as what we had
before this change, which was "if no games have been played in this
division". Players might be split into divisions mid-tourney.
Immediately after this point, Division B doesn't have any games yet.
In this case we don't want to use swissN_first_round() to generate
Division B, because its players have already played games.

swissN() should be given all games played prior to this round, even
games played when the players may have been in different divisions.

The Swiss generator is today rarely used in the first round, and
splitting players into divisions part way through the tournament is a
rubbish idea anyway, so it was unlikely this would have affected anyone.
Nevertheless, it's fixed.