Swiss generator: add option to only consider the previous round
Option to choose fitures for the next round based on a standings table
generated only from the games played in the previous round.

This is surprisingly fiddly to fit into the existing code, because all
the code in countdowntourney.py which calculates the standings uses SQL
views which fetch all games of type "P". But we now have
get_standings_from_round_onwards() which works out the "standings since
round X" by itself, which the Swiss generator uses if this option is
enabled.