Display per-player win-loss sequenceon generated fixtures and player page
When the fixture generator asks you to accept fixtures it's generated,
it is now a lot easier to see at-a-glance which players have won how
many games, and in particular, which games they were. Each player is
shown alongside their standings position, win-loss sequence (e.g. WWWL)
and win-loss record (e.g. "3-1").

This is especially useful if you're checking fixtures generated with
the "only consider the most recent round" option, because you can now
see which players won and lost the last round's games.

The player page for a specific player also shows the same information.