Add test for per-round standings function
test/standingstest.py tests the get_standings_from_round_onwards() and
get_standings() methods of the countdowntourney class, with random
results, withdrawals, and late arrivals.

Also fixed a couple of bugs in the definition of the round_standings
view and the countdowntourney.get_players_from_division() method, which
were found by the test.