Fix occasional double-wrapping of standings table column in HTML
Sometimes if a player is withdrawn, the standings table on the
public-facing display can have some HTML appear over it because
we've wrapped a field in <span class="withdrawn"> ... </span> twice.

Turns out we were passing the reference of a list rather than copying
it, and also there was a semicolon missing. Fixed.