Add "count games from round xyz" control to Standings page
The user can now view a standings table which counts only games from a
certain round onwards, if the tourney database file is new enough to
have the round_standings view.

If you use the "only consider games from the most recent round" option
to the Swiss fixture generator, this feature makes it easier to verify
that the generated fixtures make sense.