Write tourney DB version in red if older than Atropine software version
Where the sidebar shows the Atropine software version and the version of
Atropine that created this tourney's DB file, put the tourney version in
red if it's older than Atropine, with a dotted underline to draw
attention to the presence of mouseover text.

The mouseover text warns that this is an old tourney DB file and some
functionality may be unavailable.