Don't show name-to-table index if tables only have one game each
There's a new display option to "never use name-to-table index if there
are two-player tables", which defaults to true. This is because if you
have a Bristol-style event where each player plays one game in a round
and each table has one game on it, you want the display to show who's
player 1 and who's player 2.

For three-to-a-table Lincoln-style events, it doesn't matter so much
because each player takes one turn as P1 and one turn as P2.