Add run_all_tests.py to help running all tests at once
run_all_tests.py runs all the tests listed in test/tests.txt, redirects
their stdout and stderr to files under test/logs/, and reports the
outcome. These tests must all exit with status 0 to pass or nonzero
to fail.

If you're a user of Atropine and are reading this, none of it is likely
to be of any interest to you.