Pages served from webroot/cgi-bin/ are no longer run as CGI scripts
The CGIHTTPRequestHandler is deprecated in Python 3.13 and will be
removed in Python 3.15. Atropine has a lot of pages which are
generated from CGI scripts. This change makes these pages not generated
from a separate process any more, so we no longer need the
CGIHTTPRequestHandler.

All the scripts under webroot/cgi-bin are no longer standalone
"scripts" whch take their input from stdin/environment variables and
write the page to stdout. They are now modules with a handle() function,
and that handle() function contains most of the same code the main
script did before, with some necessary adaptations.

The handle() function is passed the query string, request method and
other information, along with a "response" object. Instead of writing
the HTML to standard output (cgicommon.writeln()), the function writes
the HTML to the response object (response.writeln()), and when the
handle() function returns, the caller (AtropineHTTPRequestHandler)
takes care of sending that response back to the client.

AtropineHTTPRequestHandler now subclasses SimpleHTTPRequestHandler,
not CGIHTTPRequestHandler. If it receives a request for
something matching /cgi-bin/*.py, it passes that request to the
appropriate module in webroot/cgi-bin.

The main advantage of all this, other than not using a deprecated
library, is that we no longer need to start up a whole new process for
every web page request, and the whole experience should be a bit less
sluggish. Also, we can remove a lot of repeated boilerplate code from
the individual scripts (check that the tourney name is specified, check
it exists, return an error if not, parse the query string, etc...) -
this is now done once in AtropineHTTPRequestHandler.

Quite a few pages are still a bit CGI-like in that they refresh the
page upon submitting something, but at least that no longer starts up
a new Python process.

Summary of changes:
    * Over 1,700 occurrences of cgicommon.writeln() changed to
      response.writeln().
    * cgicommon functions which "show" or "write" or "print" something
      now take a response object as the destination rather than writing
      to stdout.
    * Nothing under webroot/cgi-bin should ever be calling sys.exit()
      any more. Previously this finished the response, but now it will
      bring down Atropine.
    * Nothing under webroot/cgi-bin should be checking CGI environment
      variables like QUERY_STRING, REQUEST_METHOD etc any more.
    * Anything that was executed as part of the main script in
      webroot/cgi-bin/*.py is now in that module's handle() function.