Move webroot/cgi-bin/* to py/dynamicpages/
webroot/cgi-bin is no more. Requests for /cgi-bin/<blah>.py are now
served by calling the handle() function of the <blah> module from
py/dynamicpages/. For now, the address in the address bar will still
have cgi-bin/<blah>.py in it, but this may change soon.

Essentially, this means all the modules that were in webroot/cgi-bin
have moved to py/dynamicpages. atropine.py and
atropinehttprequesthandler.py are updated to take account of this.

The root "/" redirects to cgi-bin/home.py.