cgi-bin is no more! Long live /atropine/<tourney>/<page>/...
All pages which were previously accessed at a URL path like this:
    /cgi-bin/somescript.py?tourney=mytourney

are now here:
    /atropine/mytourney/somescript

Nothing is at /cgi-bin/ any more. All links are updated.

The service previously served from /atropine/<tourney>/state, used by the
public-facing display, is now found at /service/<tourney>/state, and
this will be true from now on for any other paths intended to be used
internally by JavaScript as opposed to being visited by the browser.

Furthermore, pages like games.py, gameslist.py and fixtureedit.py have
additional path components after the handler name. In these cases,
they are the round number. For example, the results entry page for
round 1 is now at /atropine/<tourneyname>/games/1 and not
/cgi-bin/games.py?tourney=<tourneyname>&round=1.

A new parameter extra_components is added to the former CGI scripts'
handle() function to enable this. This is a list of strings containing
the path components after the third one, i.e. the one after the handler
name. For games.py, extra_components[0] is expected to be the round
number.

fixturegen.py uses extra_components[0] for the fixture generator name,
if specified (if not specified, we display the fixgen menu).

This means we don't have to remember to pass around important and
sometimes mandatory information like the tourney name and the round
number in the query string - it's part of the URL, so forms don't
need to contain them as hidden inputs. In any case, there were many
<form> elements with unnecessary "action" attributes, where we now
omit them so the action URL is the same as the current one.

Some pages still use the query string for optional arguments, like
the tuffluck page, but for things that are always or almost always
there like the tourney name, it's neater for them to be part of the
URL path.