Better traceback pages when an uncaught exception occurs
When the pages were CGI scripts, if there was a bug and an exception was
thrown from the script, this exception could occur in the middle of the
response having been sent to the client. This meant the error message
and diagnostic information had to be displayed in an awkward pop-up div
and we had to hope the exception didn't happen in the middle of a
string, comment, <script> tag, etc.

But we don't need to do this any more. The new page modules build the
entire response before atropinehttprequesthandler sends it, so if the
module throws an exception, we can just bin off the half-built
response and return a complete page showing details of the exception
and the traceback.

So, let's do that.