Introduce "Manage Tourneys" page
Operations like downloading the db file, getting to an SQL prompt, and
now deleting the tourney can now be done from the Manage Tourneys page.
Now that the tourneys are stored in a permanent directory outside the
Atropine installation, this is an easy way to get the db files out and
delete old unneeded tourney files.

countdowntourney.py: tourney_create() has apparently always used '/'
as a directory separator, and yet this somehow always worked on Windows.
Fixed anyway.