Improve Manage Tourneys page; add functionality to import a .db file
The Manage Tourneys page now has a radio button next to each tourney,
and clicking it enables one set of buttons for things you can do to that
tourney. This looks a bit better than a massive grid of buttons, one row
for each tourney.

Tourneys can now be renamed.

Tourney .db files can now be imported into the tourney directory via the
Manage Tourneys page. To do this, we use an <input type="file" ...>
input, and for this reason the FieldStorage class has to understand
multipart/form-data. Luckily, the email module understands multipart
messages, so we can use that to parse what we get from the file upload
page.

If a request has more than 20MB of content, it is rejected with HTTP
status 413 (Content Too Large). If a POST request has no Content-Length
header, it is rejected with 411 (Length Required).

The largest POST requests we expect to receive are when someone imports
a tourney db file, but these are usually around 100KB (they're just
SQLite3 database files) so if someone's sending a request larger than
20MB they're probably doing something wrong.

The Home page and the Manage Tourneys page now have sidebars like the
other pages, albeit with far fewer links because these pages aren't
specific to one tourney.