Remove "Manage Tourneys" page and put its functionality on Home page
We don't need the recently-introduced Manage Tourneys as a separate
page. All its functionality is now on the Home page instead, and Manage
Tourneys is removed.

On the Home page, each tourney in the list now has a context menu button
to show more actions (Open, Download .db file, SQL Prompt, Rename, and
Delete). The first three are links and the last two are buttons which
show dialog boxes, but they're all styled to look like menu items.

There is now a button below the tourney list which brings up a dialog
box for importing a .db file.