Remove executable mode from most .py files
The scripts in py/dynamicpages/ used to need to be executable because
they were CGI scripts, but that's not the case any more, so they're now
non-executable.

The same goes for some files in py/ and generators/ - in the past these
may have run tests when executed as standalone scripts, but not any
more. Tests are now run from executable Python scripts in the test
directory.

Other than the test scripts, the only Python file that needs to be
executable is atropine.py.