Overhaul of Division Setup page
The Division Setup page is redone. It was initially based around the
idea that you would divide the players by rating, with players given the
option of playing in the top division if they want. This is because the
Divisions feature was initially created for a Scrabble tournament, but
this isn't how any co-event works and it's really unintuitive to use.

The concept of someone opting to be in the top division is no longer
a thing. The division_fixed column of the PLAYER table is no longer
used.

Instead, the Division Setup page now has a control to set the number of
divisions, then one table per division in which each player has a
drop-down division select box which can be changed as necessary.

If there's more than one division but all the players are in the top
division, as is the case immediately after you create additional
divisions, you get extra controls for automatically setting the
drop-down boxes based on rating, standings position, or player ID.

The number of divisions in a tournament used to be defined by the
highest-numbered division occupied by any player, which meant the last
division couldn't be empty. The number of divisions is now defined by
the "numdivisions" option in the options table, if that's greater. This
allows the Division Setup table to say "create three divisions, even if
the second and third are empty for now".

countdowntourney.set_player_divisions() is gone. This is what used to
do the automatic division assignment based on complicated rules
(optional self-promotion, number of players in a division must be a
multiple of X, etc).