Tourney uploader doesn't need to listen on localhost:3961 any more
The tourney uploader service used to listen on TCP port 3961 for
requests like "start uploading this tourney to greem", "stop uploading
this tourney", "are we uploading anything", etc. This arrangement was
required because the CGI scripts needed to talk to the uploader
service when generating pages, and these ran in their own processes.

The pages are now generated in the same process as everything else, but
they still talked to the uploader service over this TCP port, which is
a bit like two people in the same room having a conversation over the
phone. It works, but it's no longer required.

Ripped out and discarded all the TCP socket stuff between
uploadercli.py and uploader.py. uploadercli.py now just calls functions
in uploader.py directly, which feeds requests to the uploader thread.