Swiss fixgen offers to correct for latecomers' missing games
There is now a drop-down box on the Swiss fixture generator page, which
offers to make an adjustment to a player's effective standings position
if they have played fewer games than the rest.

For example, if a strong player misses round 1, normally in round 2 they
would have 0 wins and 0 points and therefore be drawn against the
lowest-scoring player(s), which wouldn't be much fun for them. So you
now have the option to count a players' missed games as either draws
(with the round's average score) or wins (with the round's highest
score) for the purpose of fixture generation.