Hidden feature to fill in unplayed games with random scores
On the results entry page, you can now enter "random scores" in both
player name boxes, and submit the form, and it will fill in any unplayed
games in that round with random scores. This is sometimes useful for
testing purposes.

The random scores are a crude simulation of a 9-round game of
Countdown, with the randomness skewed one way or the other based on
the players' ratings. It's similar to the "?showrandomresultsbutton=1"
feature the old results entry page has always had, but refined a bit.

For random scores that look more like 15-round games, put "15" in the
first score box as well as setting both player names to "random scores".