Add "Podium" section to exported results; warn if no final entered
The Text and HTML exported results now contain a "Podium" section at the
top, which lists the players in 1st, 2nd and 3rd place, taking account
of any final or semi-final/third-place playoff that has been played.

If a final has been played in a division, then 1st place goes to the
winner of that final and 2nd place goes to the loser.

If there is a third-place playoff, the winner of that is 3rd. Otherwise,
if there is no semi-final, 3rd place is whoever is 3rd in the standings.

If there is a semi-final but no third-place playoff, the Podium section
does not show who is third, because that isn't very well defined.

Additionally, if no final (game type "F") has been played, the Export
page now includes a note reminding the user of this, and that if a final
was played, they might want to create a fixture for it and enter the
result before exporting the results.