"Final" fixture generator now fails if there's already a final
The "Final between top two" fixture generator now refuses to generate a
game of type F (Final) in a division that already has one. Multiple
"Final" matches make a mess of the standings table because the code
that re-orders players to take account of the final only expects there
to be one final.

It's still not impossible to create multiple Final matches - the user
can still do it via Edit Fixtures or the Raw fixture generator, for
example - but it makes sense to at least close off the most likely route
to confusion.