Warn against renaming Prune to a new player taking Prune's place
When a new player joins mid-tournament, it is tempting for the
organiser to let them take Prune's place then rename the Automatic Prune
to the player's name. However, this is not recommended - the new player
won't appear in the standings table and the Swiss fixture generator will
always pair them with the lowest-ranked players because Atropine thinks
they're a Prune.

The "rename the Automatic Prune" control on the Advanced Setup page now
contains a note warning organisers not to do this, and instead to create
a new player and edit fixtures to reassign previous games to that player
if necessary.