Remove player "short name" attribute
If you have a tourney with a player called "Sleve McDichael" in it,
and you add a new player called just "Sleve", you get a database
exception. This is caused by setting the SHORT_NAME column of the PLAYER
table to 'Sleve' when there's already a player with this short name.

Rather than fix this bug, let's just remove the whole "short name"
feature, which has been unused for years. The "short name" is the
player's first name, or first name and initial of surname, or full
name, the shortest of these which is unique. Many years ago it was
used by an old display view that showed the standings table on the
left or right half of the screen and needed to abbreviate the player
names to do this. Nowadays it's just legacy cruft we've been carrying
around that serves no useful purpose.

Remove the SHORT_NAME column from the PLAYER table of new databases, and
remove all the associated code. In the near future it might be useful
to abbreviate a player's name to just their initials, for a wide
standings table which lists the standings after every round, but we
don't need to store a player's short name in the database.