Abbreviate player names in most exported standings progression tables
In the standings progression tables shown in the exported tournament
format, the tables for all but the first and last round show the
players' names abbreviated.

The abbreviation could be the player's initials ("FB" for Fred Bloggs)
or some extra letters to disambiguate ("FreB" and "FraB" for Fred Bloggs
and Frank Bloggs). The first and last round shows the players' full
names alongside their abbreviations.

Abbreviations are up to six letters long, plus an added number if we
still can't find an unambiguous abbreviation.

Note for the near future: currently all but the first and last round's
tables have the players' names abbreviated. This is to avoid the
text-formatted standings progression section from being too wide, but
it is probably not necessary for HTML-formatted tables. Also, even for
the text-formatted tables, we only really need to abbreviate any
players' names if the number of characters on a line would exceed a
certain amount.