Add "Progress Per Player" section to exported tournament report
The HTML and Text exported tournament reports now contain a "Progress
Per Player" section, which shows one row for each player, in
alphabetical order of name. On each row, the table shows for each round
the games that player played (in a highly abbreviated form!), and their
standings position, win count and total points (or whatever the tourney
is being ranked by) at the end of the round.

FormattedTableCell can now be given different content depending on
whether it is to be formatted as HTML or text. So when someone beat
Joe Bloggs, the text-formatted table can show a simple "W [JB]", but in
the HTML-formatted table we can put the W on a green background
(same as the fixture generator does when showing each player's form in
the generated fixtures) and show the opponent's abbreviation on a
lighter green background.

win_loss_string_to_html() still does what it did before, but its
implementation has changed to allow for optional "additional
information" next to it (the abbreviated opponent).

Progression tables are now in 10pt text rather than 12pt, to make it a
bit more likely they'll fit on the screen without needing a horizontal
scroll bar.