Progressive standings text report: only abbreviate names if > 3 rounds
The text-formatted "standings after each round" part of the exported
tournament report doesn't abbreviate the names for the first or last
round. And it seems a bit pointless to only abbreviate names for the
single middle round, so the progressive standings section (text format)
now only abbreviates the names if we're showing at least four rounds.