Introduce player rivals
New tourneys have a new database table, RIVALS. This table contains
the row (player_id, rival_id) if one of player_id's rivals is rival_id.
The rival relationship is not symmetric - player A can have player B
as a rival even if player B doesn't have player A as a rival.

A player's rivals can be set at the start of the tourney by adding one
or more "R=<name>" CSV fields to that player's line in the Tourney
Setup player list textarea, or they can be added on the player's
individual player page.

It's important that we include Rival information in the Tourney Setup
textarea, because when we submit that page, it deletes all players and
recreates them. We don't want the rival information defined elsewhere,
because then if the organiser changes a player's name in this list and
resubmits the list, any rival info will be lost.

The "rivals" field on the player page is a simple textarea containing
the names of that player's rivals, one per line.

A player cannot be a rival of themselves.

Version number bumped to 1.3.2. We only expect the RIVALS table to
exist if the DB version is >= 1.3.2.

As of this commit, player rival status doesn't actually do anything.
It is intended that matches where there exists a rival relationship
one way or the other ("grudge matches") will be highlighted, and a
new Spot Prize page can be created listing all matches where a player
beat one of their rivals.