Show "grudge matches" in remarks column on score entry screen
Press the long-neglected "Remarks" column of the game table into service
as an indicator for whether each match is a rival match for P1, for P2,
or for both.

Remove unnecessary first call to tourney.get_games() in games.py.
We never use the return value, and ater making any score updates to
the database we call tourney.get_games() again and use that.