Show rivalrous matches with an R in a circle
Fade the Remarks column back into obscurity and instead use a circle
with an R in it to denote that a player is playing a rival. This
applies to games listed on the game score entry page, the player
page, and the Defeated Rivals page.

If the game is unplayed, it's a white circle with a dotted border and
a black R in it, next to the player who is playing a rival, or both
players if they are each other's rivals. If the game is complete and
the winner beat their rival, the winner is decorated with a solid
black circle with a yellow R in it. The loser of a game does not get
any R icon.