Calculate and use lower bound on penalty for Swiss Birmingham fixgen
When equal_wins_are_equal_players is set ("Birmingham" mode), the
fixture generator can waste time trying many equivalent combinations,
swapping equivalent players, when in fact all these combinations are
optimal.

To save time when equal_wins_are_equal_players is set, work out a
lower bound on the total penalty, given the number of players in
each win group and counting how many players must be promoted to the
win group above. If in searching for a solution we find a grouping
with a penalty equal to (or less than?) this, stop searching.

In some cases this causes the Swiss fixture generator to finish quicker
in "Birmingham" mode.

Also, when we shuffle "equivalent players", leave players in win
group promotion positions where they are. We prefer to promote players
at the top of their win group, so this is likely to be a lower
penalty, and if we leave that player there, we're more likely to
find this configuration quicker rather than waiting until the search
happens to choose that player for promotion to the next win group.