Exported tourney report: highlight player name in other columns on hover
When we hover the mouse cursor over a player's name in the Progressive
Standings or Progress Per Player tables, highlight that cell and any
other cell which refers to that (possibly abbreviated) player name.

We do this by attaching the HTML class "highlightonhover" and the
attribute data-hover-key="<player name>" to all elements which
contain a player name and we want to be highlighted. All matching
elements are highlighted within the same boundary element. A boundary
element is one with "highlightonhoverboundary" in the class list.

This makes it easier to see at a glance how John Smith progressed
through the tourney, by hovering over one of the John Smith cells in
the "Standings after each round" section.

Remove the duplicate definition of is_draw() in the class Game in
countdowntourney.py. Duplicate definitions of methods in Python classes
don't produce an error or warning, the second definition just silently
replaces the first. Who knew?