Improve Round Robin fixture generator
The Round Robin fixture generator now assigns initial positions to the
players randomly, rather than having it based on the alphabetical order
of the players' names.

It also now changes the P1/P2 order so that a player in a round robin
does not play on the same side of the scoresheet more than twice in a
row. Previously we would have a player who was, say, P2 in their first
three games then P1 in their last four.

Add a test for the round robin generator. There wasn't one before.