Table numbers can now start from a number greater than 1
The Tourney Setup page now has a control for setting the first table
number. This is almost always 1, but you might have two tournaments
going on in the same room, with the first few tables reserved for the
other tournament.

The first table number must be a positive integer. If we ever wanted to
allow a Table 0, the Player page would have to find some other way to
express that a player has no specific table preference - currently this
is represented by a 0 in that field.