Improve progress-by-player table in exported tournament report
Get rid of the compact name-abbreviating table, in favour of a table
that take up a bit more space but are a lot clearer.

The "export as text" function no longer includes the "standings after
each round" section. It's too difficult to display this in a way that
won't get mangled if the spacing is messed about with (and C4C will
mess about with the spacing). So this is now only in the HTML version.

The "export as HTML" function now displays the "progress by player"
section as a larger table, with one column per round, and one set of
rows per player. Within each player row, there's a row for each game
that player played in a round, and another row for their standings
information at the end of that round. There are also no player name
abbreviations.

Also some miscellaneous fixes, including HTML-escaping things that
should be HTML-escaped, like player names.